Use typecasting to differentiate listed bands on splits
Because strings are used to identify both listed *and* unlisted bands on
a split's tracks, it becomes impossible to tell them apart later:

    Listed Band:    #2422        -> "2422"
    Unlisted Band:  "Ahasverus"  -> "Ahasverus"

This is problematic if an unlisted band uses a number for a name:

    Listed:    #2422  -> "2422"
    Unlisted:  "420"  -> "420"

To circumvent this, the IDs of listed bands are now stored as integers:

    Listed:    #2422  -> 2422
    Unlisted:  "420"  -> "420"

Storing band references as objects would be easier and clearer, but this
would be overkill considering splits with unlisted bands are a minority.
It's also important to keep the size of the exported data as light as we
can make it - some exports can potentially be very big.