Allow unlisted bands to be properly credited on splits
Sometime in mid-2015, the ability to assign a member to an unlisted band
was revoked. This was done to prevent artists with zero metal connection
from being added to the database.

Even though most splits have had unlisted bands purged from their lineup
info since then, some releases still retain this information. Since we'd
rather not erroneously list an artist as having played with another band
on a release, we need to check for this.