Improve resilience against remote disconnection
Requesting too many resources within a single session invokes connection
issues with MA's server. This is especially problematic for medium-sized
or large bands that require many hundreds of HTTP requests to load.

This commit instructs the program to reattempt any HTTP requests that've
been reset by the server. This only solves half the problem: the program
hangs after loading images if there've been connection difficulties, and
there appears to be some discrepancy in the data that's loaded between 2
different exports for the same band.

Furthermore, the handling of JSON data needs improved error-handling. If
the server returns an internal error, a regular HTML document describing
the fault is loaded instead of JSON. This needs to be taken into account
when improving the program's resilience against load failure.