Avoid overwriting previously-loaded Edit details
Some Edit instances are shared between resources. For example, an artist
added to a release's recording line-up will have an Edit instance that's
listed in the modification histories of all three objects (band, release
and the artist themselves). We need to check if an Edit object's details
have been previously loaded so we don't clobber them.